pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kimi-cli"
version = "0.1.0"
description = "AI-powered PowerPoint presentation generator"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Kimi CLI Team", email = "team@kimi-cli.dev"}
]
keywords = ["ai", "powerpoint", "presentation", "cli", "typer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Presentation",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]

dependencies = [
    "typer[all]>=0.9.0",
    "python-pptx>=0.6.21",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/lsunay/ai-template-slides"
Repository = "https://github.com/lsunay/ai-template-slides"
Issues = "https://github.com/lsunay/ai-template-slides/issues"

[project.scripts]
kimi-cli = "kimi_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["kimi_cli*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
kimi_cli = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=kimi_cli --cov-report=term-missing --cov-report=html"